validaciones por los cambios de cristian en la tabla deaths

+++ db/migrate/20221118131327_add_deaths_changes.rb
@@ -1,9 +1,9 @@
 class AddDeathsChanges < ActiveRecord::Migration[6.1]
   def change
-    remove_column :deaths, :assasin
-    remove_column :deaths, :victim
-    add_column :deaths, :assasin_id, :integer, null: false
-    add_column :deaths, :victim_id, :integer, null: false
+    remove_column :deaths, :assasin if ActiveRecord::Base.connection.column_exists?(:deaths, :assasin)
+    remove_column :deaths, :victim if ActiveRecord::Base.connection.column_exists?(:deaths, :victim)
+    add_column :deaths, :assasin_id, :integer, null: false unless ActiveRecord::Base.connection.column_exists?(:deaths, :assasin_id)
+    add_column :deaths, :victim_id, :integer, null: false unless ActiveRecord::Base.connection.column_exists?(:deaths, :victim_id)
     add_column :deaths, :first_blood, :boolean, default: false
   end
 end
